temporarily move logout functionality to logout fragment
